test: align substring i64::MAX literal with i32::MAX and drop stale aggregate-MV rejection
cargo test was failing in 4 places under the new SUBSTR / managed-lake MV semantics:

1. `tests/function/string.rs` — three substr tests passed `i64::MAX` as the
   length literal. PR #119 added a runtime narrowing that turns column-supplied
   BIGINT overflow into NULL (mirrors StarRocks FE's `BIGINT → INT` cast); but
   the runtime also uses `i64::MAX` as the internal "no length" sentinel for
   the 2-arg form (see `src/exec/expr/function/string/substring.rs` line 232).
   When the literal `i64::MAX` reaches the runtime via these unit tests
   (which bypass the analyzer), the narrowing nukes it instead of letting the
   sentinel handler return the rest of the string. Switch to `i32::MAX as i64`
   to match StarRocks BE's own substr tests (which use `INT_MAX`, not
   `LONG_MAX`).

2. `tests/standalone_mysql_server.rs` — drop
   `standalone_mysql_server_mv_incremental_rejects_aggregate_definition`. The
   test asserted that aggregate-defined incremental MVs are rejected with
   "projection/filter", but PR #62 added `AggregateMvShape` and `JoinAggregate`
   support, so that contract no longer holds. The test panicking inside a
   static `Mutex` ServerGuard poisoned the lock for every other parallel
   `standalone_mysql_server_*` test, surfacing as 15 collateral PoisonError
   failures.

3. `tests/standalone_mysql_server.rs::mv_show_output_matches_expected_columns`
   — `SHOW MATERIALIZED VIEWS` gained a `StorageEngine` column between
   `Database` and `RefreshMode` after PR #133/#143; update the 7-tuple row
   type to an 8-tuple, add the `managed_lake` assertion, and shift the index
   labels accordingly.

`cargo test`: 3104+ passed / 0 failed / 3 ignored.

+++ tests/function/string.rs
@@ -1644,9 +1644,11 @@ fn test_substring_multiple_rows() {
 fn test_substring_edge_cases() {
     let mut arena = ExprArena::default();
 
-    // Edge cases from StarRocks test suite
-    test_substring_3args(&mut arena, "123456789", 1, i64::MAX, "123456789");
-    test_substring_3args(&mut arena, "123456789", 2, i64::MAX, "23456789");
+    // Edge cases from StarRocks test suite. StarRocks BE uses INT_MAX (i32::MAX)
+    // because the BE substr signature narrows BIGINT→INT; passing i64::MAX here
+    // would collide with the runtime sentinel reserved for the 2-arg form.
+    test_substring_3args(&mut arena, "123456789", 1, i32::MAX as i64, "123456789");
+    test_substring_3args(&mut arena, "123456789", 2, i32::MAX as i64, "23456789");
     test_substring_3args(&mut arena, "123456789", -9, 9, "123456789");
     test_substring_3args(&mut arena, "123456789", -9, 10, "123456789");
     test_substring_3args(&mut arena, "123456789", -4, 5, "6789");
@@ -1704,7 +1706,7 @@ fn test_substring_comprehensive_starrocks_cases() {
     test_substring_3args(&mut arena, "123456789", -1, 1, "9");
     test_substring_3args(&mut arena, "123456789", -1, 2, "9");
     test_substring_3args(&mut arena, "123456789", 0, 1, "");
-    test_substring_3args(&mut arena, "123456789", 1, i64::MAX, "123456789");
+    test_substring_3args(&mut arena, "123456789", 1, i32::MAX as i64, "123456789");
     test_substring_3args(&mut arena, "123456789", 1, -2, "");
     test_substring_3args(&mut arena, "123456789", -3, -2, "");
 }
@@ -1731,7 +1733,13 @@ fn test_substring_chinese_comprehensive() {
     test_substring_3args(&mut arena, test_str, -1, 1, "玖");
     test_substring_3args(&mut arena, test_str, -1, 2, "玖");
     test_substring_3args(&mut arena, test_str, 0, 1, "");
-    test_substring_3args(&mut arena, test_str, 1, i64::MAX, "壹贰叁肆伍陆柒捌玖");
+    test_substring_3args(
+        &mut arena,
+        test_str,
+        1,
+        i32::MAX as i64,
+        "壹贰叁肆伍陆柒捌玖",
+    );
 }
 
 // ---------------------------------------------------------------------------
